+++ config/environment.js
@@ -8,9 +8,7 @@ nconf
 
 nconf.defaults({
   'ENVIRONMENT': 'production',
-  'RIPPLE_REST_API': 'http://localhost:5990/',
-  'RIPPLE_REST_PATH': null,
-  'RUN_RIPPLE_REST': false,
+  'RIPPLE_REST_API': 'https://api.ripple.com/',
   'DATABASE_USER': 'postgres',
   'DATABASE_PASSWORD': 'password',
   'DATABASE_NAME': 'ripple_gateway',
@@ -21,7 +19,6 @@ nconf.defaults({
   //
   // DEPRECATED: Use DATABASE_{USER|PASSWORD|...} instead.
   'DATABASE_URL': null,
-  'RIPPLE_DATAMODEL_ADAPTER': 'ripple-gateway-data-sequelize',
   'SSL': true,
   'SSL_KEY_PATH': __dirname+'/../env/certs/server.key',
   'SSL_CERTIFICATE_PATH': __dirname+'/../env/certs/server.crt',
@@ -39,10 +36,6 @@ nconf.defaults({
   'CURRENCIES': null, // Required
   'WITHDRAWAL_FEE': 0.01, // Required - default 1%
   'DEPOSIT_FEE': 0.01, // Required - default 1%
-  'WITHDRAWALS_CALLBACK_URL': null, // URL to POST callback of withdrawals
-  'WITHDRAWALS_CALLBACK_CERTIFICATE': null, // Used to validate SSL on callbacks
-  'DEPOSITS_CALLBACK_URL': null, // URL to POST callback of withdrawals
-  'DEPOSITS_CALLBACK_CERTIFICATE': null, // Used to validate SSL on callbacks
   'LOGGLY': false
 });