+++ index.js
@@ -10,6 +10,13 @@ exports.ripple = require(__dirname +'/lib/ripple/');
 exports.validator = require(__dirname+'/lib/validator.js');
 exports.server = require(__dirname+'/lib/app.js');
 exports.processes = require(__dirname+'/lib/processes/');
+exports.logger = require(__dirname+'/lib/data/logs.js');
+
+var initializers = requireAll(__dirname+'/config/initializers/');
+
+for (i in initializers) {
+  initializers[i](exports);
+}
 
 require(__dirname+'/Gatewaydfile')(exports);