+++ processes/incoming.js
@@ -1,12 +1,21 @@
-var gateway = require(__dirname+'/../');
+var Promise = require('bluebird');
+Error.stackTraceLimit = Infinity;
 
+process.on('uncaughtException', function(err) {
+  console.log('Caught exception: ' + err);
+});
+
+Promise.onPossiblyUnhandledRejection(function(err) {
+  console.log('Caught exception: ' + err);
+});
+
+var gatewayd = require(__dirname+'/../');
 var Listener = require(__dirname+'/../lib/ripple/listener.js');
-
 var listener = new Listener();
 
 listener.onPayment = function(payment) {
-  logger.info('payment:notification:received', payment);
-  if (payment && payment.destination_account === gateway.config.get('COLD_WALLET')) {
+  gatewayd.logger.info('payment:notification:received', payment);
+  if (payment && payment.destination_account === gatewayd.config.get('COLD_WALLET')) {
     var opts = {
       destinationTag : payment.destination_tag,
       transaction_state : payment.result,
@@ -20,14 +29,14 @@ listener.onPayment = function(payment) {
           opts.currency = balanceChange.currency;
           opts.issuer = balanceChange.issuer;
           opts.state = 'incoming';
-          gateway.api.recordIncomingPayment(opts, function(error, record) {
+          gatewayd.api.recordIncomingPayment(opts, function(error, record) {
             if (error) {
-              logger.error('payment:incoming:error', error);
+              gatewayd.logger.error('payment:incoming:error', error);
             } else {
               try {
-                logger.info('payment:incoming:recorded', record.toJSON());
+                gatewayd.logger.info('payment:incoming:recorded', record.toJSON());
               } catch(exception) {
-                logger.error('payment:incoming:error', exception);
+                gatewayd.logger.error('payment:incoming:error', exception);
               }
             }
           });
@@ -37,19 +46,27 @@ listener.onPayment = function(payment) {
   }
 };
 
-var lastHash = gateway.config.get('LAST_PAYMENT_HASH');
+var lastHash = gatewayd.config.get('LAST_PAYMENT_HASH');
 
-if (lastHash) {
-  listener.start(lastHash);
-  logger.info('Listening for incoming ripple payments from Ripple REST, starting at', lastHash);
-} else {
-  console.log('LAST_PAYMENT_HASH not set... gatewayd is now fetching it from Ripple.');
-  gateway.api.fetchLastPaymentHash().then(function(hash) {
-    gateway.config.set('LAST_PAYMENT_HASH', hash)
-    gateway.config.save(function() {
-      listener.start(hash);
-      logger.info('Listening for incoming ripple payments from Ripple REST, starting at', hash);
+gatewayd.api.getColdWallet(function(error, address) {
+  if (error) {
+    throw new Error(error);
+  }
+  if (!address) {
+    throw new Error('Ripple COLD_WALLET not set');
+  }
+  if (address.getLastPaymentHash()) {
+    const hash = address.getLastPaymentHash();
+    listener.start(hash); 
+    logger.info('Listening for incoming ripple payments from Ripple REST, starting at', hash);
+  } else {
+    logger.info('LAST_PAYMENT_HASH not set... gatewayd is now fetching it from Ripple.');
+    address.fetchLastPaymentHash().then(function(hash) {
+      address.setLastPaymentHash(hash).then(function() {
+        listener.start(hash);
+        logger.info('Listening for incoming ripple payments from Ripple REST, starting at', hash);
+      });
     });
-  });
-}
+  }
+});