+++ test/core/outgoing_payment.js
@@ -1,126 +1,147 @@
 var assert = require('assert');
 var gateway = require(__dirname+'/../../');
 var OutgoingPayment = require(__dirname+'/../../lib/core/outgoing_payment.js');
+var fixtures = require(__dirname+'/../fixtures/ripple_rest_integration.js');
 var outgoingTransactionRecord;
-var restConfirmationResponse;
+
 
 describe('Outgoing Payment', function() {
-  describe('Sending a queued payment to ripple', function() {
+  describe('Ripple REST error handler', function(){
 
-    it('should mark a record as sent once successfully communicated to ripple rest', function(done) {
-      var outgoingPayment = new OutgoingPayment(outgoingTransactionRecord);
-      assert.strictEqual(outgoingPayment.record.state, 'outgoing');
+    describe('transactional errors - fund error', function(){
+      before(setUp);
+      after(tearDown);
 
-      var rippleRestResponse = {
-        success: true,
-        client_resource_id: 'a2837056-20e8-410a-b730-879dd0493742',
-        status_url: 'http://localhost:5990/v1/accounts/rscJF4TWS2jBe43MvUomTtCcyrbtTRMSNr/payments/a2837056-20e8-410a-b730-879dd0493742'
-      };
+      it('should handle insufficient fund error and mark record failed', function(done) {
+        var outgoingPayment = new OutgoingPayment(outgoingTransactionRecord);
+        outgoingPayment._rippleRestResponseHandler(fixtures.errors.insufficient_funds)
+          .then(function(error){
+            assert.strictEqual(error.record.state, 'failed');
+            assert.strictEqual(error.record.transaction_state, 'terINSUF_FEE_B');
+            done();
+          });
+      });
 
-      outgoingPayment._markRecordAsSent(rippleRestResponse, function() {
-        assert.strictEqual(outgoingPayment.record.state, 'sent');
-        done();
+      it('should handle transaction not found -- retry', function(done) {
+        var outgoingPayment = new OutgoingPayment(outgoingTransactionRecord);
+        outgoingPayment._rippleRestResponseHandler(fixtures.errors.invalid_requests.transaction_not_found)
+          .then(function(error){
+            assert.deepEqual(error.record, outgoingTransactionRecord.dataValues);
+            assert.strictEqual(error.handled, 'retry');
+            done();
+          });
+
       });
     });
 
-    it('should confirm the acceptance of the payment into the Ripple Ledger', function(done) {
-      var outgoingPayment = new OutgoingPayment(outgoingTransactionRecord);
-      var rippleRestResponse = {
-        success: true,
-        client_resource_id: 'a2837056-20e8-410a-b730-879dd0493742',
-        status_url: 'http://localhost:5990/v1/accounts/rscJF4TWS2jBe43MvUomTtCcyrbtTRMSNr/payments/a2837056-20e8-410a-b730-879dd0493742'
-      };
+    describe('transactional errors -- no path found', function() {
 
-      outgoingPayment._confirmSuccessfulPayment(rippleRestResponse, function(error) {
-        assert.strictEqual(outgoingPayment.record.state, 'sent');
-        assert(!error);
-        done();
+      setTimeout(function(){
+        before(setUp);
+        after(tearDown);
+        it('should handle no path found -- failed', function(done) {
+          this.timeout(5000);
+          var outgoingPayment = new OutgoingPayment(outgoingTransactionRecord);
+          outgoingPayment._rippleRestResponseHandler(fixtures.errors.invalid_requests.no_paths_found)
+            .then(function(error){
+              assert.strictEqual(error.record.transaction_state, 'tecPATH_DRY');
+              assert.strictEqual(error.handled, 'failed');
+              done();
+            });
+        });
+      }, 5000);
+    });
+
+    describe('connection errors', function(){
+      it('should handle no rippled connection error -- retry', function(done) {
+        var outgoingPayment = new OutgoingPayment(outgoingTransactionRecord);
+        outgoingPayment._rippleRestResponseHandler(fixtures.errors.connection.no_rippled_connection)
+          .then(function(error){
+            assert.deepEqual(error.record, outgoingTransactionRecord.dataValues);
+            assert.strictEqual(error.handled, 'retry');
+            done();
+          });
       });
-    });
 
-    it('should record the rejection of a payment from the Ripple Ledger', function(done) {
-      var outgoingPayment = new OutgoingPayment(outgoingTransactionRecord);
-      var confirmationResponse = new Object(restConfirmationResponse);
-      confirmationResponse.result = 'temPATH_DRY';
-
-      outgoingPayment._recordAcceptanceOrRejectionStatus(confirmationResponse, function(error) {
-        assert(!error);
-        assert.strictEqual(outgoingPayment.record.state, 'failed');
-        done();
+      it('should handle rippled busy error -- retry', function(done) {
+        var outgoingPayment = new OutgoingPayment(outgoingTransactionRecord);
+        outgoingPayment._rippleRestResponseHandler(fixtures.errors.connection.rippled_busy)
+          .then(function(error){
+            assert.deepEqual(error.record, outgoingTransactionRecord.dataValues);
+            assert.strictEqual(error.handled, 'retry');
+            done();
+          });
       });
-    });
 
-    it('should record the failure of a payment upon submission to ripple rest', function(done) {
-      var outgoingPayment = new OutgoingPayment(outgoingTransactionRecord);
-      outgoingPayment._handleRippleRestFailure('No path found. Please try ....', function(error) {
-        assert(!error);
-        assert.strictEqual(outgoingPayment.record.state, 'failed');
-        done();
+      it('should handle socket hang up connection error -- retry', function(done) {
+        var error = new Error('socket hang up');
+        error.code = 'ECONNRESET';
+
+        var outgoingPayment = new OutgoingPayment(outgoingTransactionRecord);
+        outgoingPayment._rippleRestResponseHandler(error)
+          .then(function(error){
+            assert.deepEqual(error.record, outgoingTransactionRecord.dataValues);
+            assert.strictEqual(error.handled, 'retry');
+            done();
+          });
+
       });
+
     });
 
-    it('should retry a payment by moving it back to the "outgoing" queue', function(done) {
-      var outgoingPayment = new OutgoingPayment(outgoingTransactionRecord);
-      outgoingPayment._handleRippleRestFailure('retry', function(error) {
-        assert(!error);
-        assert.strictEqual(outgoingPayment.record.state, 'outgoing');
-        done();
-      });
-    });
-
-    var outgoingPayment;
-
-    before(function(done) {
-      outgoingPayment = new OutgoingPayment(outgoingTransactionRecord);
-      done();
-    });
-
-  });
-
-  before(function(done) {
-    gateway.data.models.rippleTransactions.create({
-      to_address_id: 820,
-      from_address_id: 623,
-      to_amount: '0.0019399999999999999',
-      to_currency: 'XRP',
-      to_issuer: 'r4EwBWxrx5HxYRyisfGzMto3AT8FZiYdWk',
-      from_amount: '0.002',
-      from_currency: 'XRP',
-      from_issuer: 'r4EwBWxrx5HxYRyisfGzMto3AT8FZiYdWk',
-      state: 'outgoing',
-      external_transaction_id: 169
-    }).complete(function(error, rippleTransaction) {
-      outgoingTransactionRecord = rippleTransaction;
-      done();
-    });
-
-    restConfirmationResponse = {
-      source_account: 'rscJF4TWS2jBe43MvUomTtCcyrbtTRMSNr',
-      source_tag: '623',
-      source_amount: { value: '0.001939', currency: 'XRP', issuer: '' },
-      source_slippage: '0',
-      destination_account: 'rp4u5gEskM8DtBZvonZwbu6dspgVdeAGM6',
-      destination_tag: '',
-      destination_amount: { value: '0.001939', currency: 'XRP', issuer: '' },
-      invoice_id: '',
-      paths: '[]',
-      no_direct_ripple: false,
-      partial_payment: false,
-      direction: 'outgoing',
-      state: 'validated',
-      result: 'tesSUCCESS',
-      ledger: '',
-      hash: '9DDCEBAB6D751C22755F9303B59E1E2FDC8308B551A4C5AE89343BD6F2255169',
-      timestamp: '2014-06-30T00:28:46.000Z',
-      fee: '0.000012',
-      source_balance_changes: [],
-      destination_balance_changes: []
-    };
-
-  });
-
-  after(function(done) {
-    outgoingTransactionRecord.destroy().complete(done);
   });
 
 });
+
+
+describe('Sending a queued payment to ripple', function() {
+
+  before(setUp);
+  after(tearDown);
+
+  it('should record the rejection of a payment from the Ripple Ledger', function(done) {
+    var outgoingPayment = new OutgoingPayment(outgoingTransactionRecord);
+    var confirmationResponse = new Object(fixtures.successful_responses.validated_payment);
+    confirmationResponse.result = 'temPATH_DRY';
+
+    outgoingPayment._recordAcceptanceOrRejectionStatus(confirmationResponse)
+      .then(function(status) {
+        assert.strictEqual(outgoingPayment.record.state, 'failed');
+        done();
+      });
+  });
+
+  it('should record sent and confirmed a ripple payment with a hash', function(done){
+    this.timeout(10000);
+    var outgoingPayment = new OutgoingPayment(outgoingTransactionRecord);
+    outgoingPayment._sendAndConfirmPayment(fixtures.requests.payment)
+      .then(function(payment){
+        assert(payment.hash);
+        assert.strictEqual(payment.result, 'tesSUCCESS');
+        done();
+      });
+  });
+
+});
+
+function setUp(done) {
+  gateway.data.models.rippleTransactions.create({
+    to_address_id: 820,
+    from_address_id: 623,
+    to_amount: '0.0019399999999999999',
+    to_currency: 'XRP',
+    to_issuer: 'r4EwBWxrx5HxYRyisfGzMto3AT8FZiYdWk',
+    from_amount: '0.002',
+    from_currency: 'XRP',
+    from_issuer: 'r4EwBWxrx5HxYRyisfGzMto3AT8FZiYdWk',
+    state: 'outgoing',
+    external_transaction_id: 169
+  }).complete(function (error, rippleTransaction) {
+    outgoingTransactionRecord = rippleTransaction;
+    done();
+  });
+}
+
+function tearDown(done) {
+  outgoingTransactionRecord.destroy().complete(done);
+}